feat: 컴포넌트 분리 및 행사 타입 칩 추가
