+++ src/components/chips/Chip.tsx
@@ -13,8 +13,8 @@ const Chip = ({ children, onClick, isSelected }: Props) => {
       onClick={onClick}
       className={`inline-flex items-center justify-center whitespace-nowrap break-keep rounded-[28px] px-12 py-8 text-14 font-600 leading-[160%] ${
         isSelected
-          ? 'bg-brand-primary-400 text-basic-white active:bg-brand-primary-500'
-          : 'bg-basic-grey-100 text-basic-grey-700 active:bg-basic-grey-200'
+          ? 'bg-brand-primary-400 text-basic-white'
+          : 'bg-basic-grey-100 text-basic-grey-700'
       }`}
     >
       {children}
